Changed duration from double to int

+++ src/Modules/BusinessProfile/Servibes.BusinessProfile.Api/Commands/Services/ServiceDto.cs
@@ -6,7 +6,7 @@ namespace Servibes.BusinessProfile.Api.Commands.Services
     {
         public string ServiceName { get; set; }
         public decimal Price { get; set; }
-        public double Duration { get; set; }
+        public int Duration { get; set; }
         public string Description { get; set; }
         public List<EmployeeDto> Performers { get; set; }
     }
